requirements.txt
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
django-extensions==3.0.7
django-filter==2.4.0
django-heroku==0.3.1
filelock==3.0.12
graphviz==0.14.1
gunicorn==20.0.4
Markdown==3.3.3
Pillow==8.0.1
pipenv==2020.11.15
psycopg2==2.8.6
python-decouple==3.3
requests==2.25.1
whitenoise==5.2.0
qrcode
django-social-share
